test(endpoint-share): use node test runner
