--- packages/endpoint-share/tests/integration/302-post-share.js
@@ -1,24 +1,0 @@
-import test from "ava";
-import supertest from "supertest";
-import { mockAgent } from "@indiekit-test/mock-agent";
-import { testServer } from "@indiekit-test/server";
-import { testCookie } from "@indiekit-test/session";
-import { testToken } from "@indiekit-test/token";
-
-await mockAgent("endpoint-micropub");
-
-test("Posts content and redirects back to share page", async (t) => {
-  const server = await testServer();
-  const request = supertest.agent(server);
-  const result = await request
-    .post("/share")
-    .set("cookie", [testCookie()])
-    .send(`access_token=${testToken()}`)
-    .send("name=Foobar")
-    .send("content=Test+of+sharing+a+bookmark")
-    .send("bookmark-of=https://example.website");
-
-  t.is(result.status, 302);
-
-  server.close(t);
-});
